refactor and break down into smaller files
