+++ birdy/exceptions.py
@@ -1,5 +1,10 @@
 import click
+from owslib.util import ServiceException
 
 
 class ConnectionError(click.ClickException):
+    pass
+
+
+class UnauthorizedException(ServiceException):
     pass
