+++ birdy/utils.py
@@ -1,6 +1,3 @@
-import collections
-
-import six
 from six.moves.urllib.parse import urlparse
 
 
@@ -12,13 +9,3 @@ def is_url(url):
         return False
     else:
         return True
-
-
-def delist(data):
-    if (
-        isinstance(data, collections.Iterable)
-        and not isinstance(data, six.string_types)
-        and len(data) == 1
-    ):
-        return data[0]
-    return data
